--- assets/src/domain/eventEditor/ui/datetimes/dateCard/style.css
@@ -1,3 +1,0 @@
-.bp3-popover-content .ee-date-range-picker {
-	padding: 1rem;
-}
