--- assets/src/domain/eventEditor/ui/datetimes/dateForm/style.ts
@@ -1,42 +1,0 @@
-import { CSSProperties } from 'react';
-
-export const hdrStyle: CSSProperties = {
-	margin: '1em 0 .5em 24%',
-};
-
-export const lblStyle: CSSProperties = {
-	boxSizing: 'border-box',
-	color: 'grey',
-	display: 'inline-block',
-	fontSize: '1em',
-	lineHeight: '2rem',
-	minWidth: '80px',
-	paddingRight: '1em',
-	textAlign: 'right',
-	verticalAlign: 'top',
-	width: '24%',
-};
-
-export const inputStyle: CSSProperties = {
-	boxSizing: 'border-box',
-	display: 'inline-block',
-	fontSize: '1em',
-	lineHeight: '2rem',
-	minWidth: '200px',
-	width: '60%',
-};
-
-export const divStyle: CSSProperties = {
-	boxSizing: 'border-box',
-	display: 'block',
-	margin: '0 0 1em',
-	width: '100%',
-};
-
-export const relationsStyle: CSSProperties = {
-	boxSizing: 'border-box',
-	display: 'inline-block',
-	marginLeft: '24%',
-	padding: '.5em 0 1em',
-	width: '60%',
-};
