+++ assets/src/domain/eventEditor/ui/datetimes/datesList/actionsMenu/EditDateButton.tsx
@@ -3,9 +3,9 @@ import { EditOutlined } from '@ant-design/icons';
 import { __ } from '@wordpress/i18n';
 
 import { EspressoButton } from '@application/ui/input';
-import { EditItemButtonProps } from '../../../interfaces/types';
-import { useDatetimeContext } from '../../../hooks';
-import { useEditorModal } from '../../../../../application/ui/layout/editorModal';
+import { EditItemButtonProps } from '@edtrInterfaces/types';
+import { useDatetimeContext } from '@edtrHooks/index';
+import { useEditorModal } from '@appLayout/editorModal';
 
 const EditDateButton: React.FC<EditItemButtonProps> = ({ ...rest }) => {
 	const { id: entityId } = useDatetimeContext();
