+++ assets/src/domain/eventEditor/ui/datetimes/datesList/tableView/DatesListView.js
