+++ assets/src/domain/eventEditor/ui/datetimes/datesList/tableView/editor-date-entities-list-view.css
